requirements.txt
linkedin-api-client
python-dotenv
playwright